Add abortwrite() and move abortRead().
Soon, shutdownWrite() will go away, replaced with a way to send clean EOF that flushes buffers and returns a promise.

However, sometimes you want to cancel writes without waiting. The new abortWrite() method serves that purpose, and is analogous to the existing abortRead().

This change also moves these methods to be on the respective input/output stream types rather than on the combined AsyncIoStream. Historically I put shutdownWrite() and abortRead() on AsyncIoStream because I thought they were redundant on one-way streams since you could just delete the stream and let the destructor do this. This was a mistake and I'm now trying to separate these functions from destruction.