Fix bug with tail calls and pipelining.
The destructor of RpcCallContext would prematurely free the pipeline object in the case of tail calls, because it treated them the same as cancellation. When a call is canceled by the client (by an early `Finish` message), we know that the client won't make any pipelined calls on it. But in the case of a tail call, RpcCallContext's destructor may be invoked before the client has sent a `Finish`. The client might still send pipelined calls in this case!