Remove `UnalignedFlatArrayMessageReader`.
This reverts commit 70b1e1ff10ae58b908f902013d15369deee6ad82.

This is not safe, even on x86. Unaligned access is undefined behavior in C/C++. Compilers can and do optimize under the assumption that access is aligned. For example, GCC may use SIMD instructions in optimizations, and these do in fact require alignment. In principle, a compiler could also choose to mask away the lowest bits of a pointer that should always be zero when aligned.