Fix -fno-exceptions build.
Note that executor disconnects still aren't super-usable when exceptions are disabled as trying to schedule on an executor whose thread is _already_ gone throws a fatal exception. I don't care to put in the effort to fix this because I don't think anyone will use this feature with exceptions disabled. TBH, we should probably define a small subset of KJ/capnp that are expected to work with exceptions disabled and stop trying to test the rest in this configuration...