Actually fix windows.
Duh, we need to call ensureThreadCanRunFibers() during runSynchronously(). I "accidentally" fixed it before by making it so EventLoop didn't un-fiber the thread on shutdown, which mean that subsequent tests calling runSynchronously() worked -- the runSynchronously() test would have failed if executed in isolation. But now we have a test that spawns a thread and executes runSynchronously() there, and that definitely failed.