Allow MembranePolicies to opt out of resolve-before-redirect behavior.
Historically, if a MembranePolicy decides to intercept a call, but the original call target was an unresolved promise capability, then the membrane would discard the redirect and instead wait for the capability to resolve and then apply the policy again.

This behavior is important for reversible MembranePolicies -- that is, policies where it's important that a capability passed over the membrane _and then passed back_ ends up having the membrane removed. With such policies, it could turn out that a promise pointing into the membrane eventually resolves to an object outside the membrane. In that case, the policy should not apply after all. It would be confusing if the policy applies to a promise but not to its resolution -- this could create a difference in behavior based on timing of a call.

However, some MembranePolicies don't actually care about reversibility, and in these cases, forcing resolution of promises before the policy can apply may add undesirable delays.

Therefore, this PR makes the behavior configurable, with yet another virtual method.