Fix TextCodec parse error formatting.
The error messages always started with `kj::str(message, " (", startByte, ":", endByte, ").") =` due to the use of KJ_FAIL_ASSERT. This garbage clearly wasn't intended to be part of the error message.

I also went a little further and computed the line number and column numbers, which is usually nicer than a byte span.