Redesign async trace code.
The old Event::trace() in particular was nonsense -- it didn't actually trace back to the next event and so on, instead it would trace _forwards_ through `PromiseNode`s.

Additionally, the output of the old traces was not particularly useful, as it would tell you the types of the promises but not line numbers.

The new tracing code produces a list of instruction addresses, just like regular stack traces. It can be passed into all the regular stack tracing tools.

Additionally, a new global function `kj::getAsyncTrace()` gets a trace of the promises leading up to the currently-executing code. Previously, such traces were generated when you threw an exception and it actually passed through the promise chain, but there was no way to get a meaningful async trace e.g. in a crash handler. Now there is.