Re-arrange QuestionRef destructor to handle exceptions better.
`message->send()` could throw for all sorts of reasons. Previously, in this case, we failed to null out `question.selfRef`, leaving it dangling.

In fact, we probably don't want to propagate these exceptions out of `~QuestionRef()` -- it is refcounted, so this could throw at a weird spot. Instead, we should proabbly kill the connection, by catching and calling `disconnect()`.