Remove a redundant static_cast.
This `static_cast` is casting the value to the type it already is. I think it may have been a copy-pasto from a version of the code that actually had to do a downcast here.

This is just something I noticed while reading. This change has absolutely no effect on actual behavior.