Delete sysconf-based iovMax() fallback.
Apparently, no platform that we target today needs this, and as a result it's totally untested, which probably means it doesn't work anyway. I think I copy/pasted this fallback code from a stack overflow or something a very long time ago.

If, in the future, we encounter a platform that doesn't have one of the constants defined, then we should probably either figure out if that platform provides some other constant, or just determine a number that works and hard-code it.