AsyncPipe pumps should fail when any read/write fails.
When an AsyncPipe's read end is pumping to some other stream, and a write is performed on the write end, the write propagates through to the target stream. If that write fails, previously the exception would only propagate to the caller of write(). The pump would stay in place. This is probalby undesirable because it's probably not possible to write through the pump any more without corruption, and anyway it doesn't match the behavior of the generit pumpTo() implementation. This change propagates the exception from the write to the pumpTo() operation.

A similar argument seems to apply to all other permutations of operations as well, so I updated all of them, though the above scenario is the ony one that was explicitly causing a problem for me.