Fix bug when exporting a promise capability twice.
If the same promise capability was exported more than once -- reusing the same export table entry -- it was not marked as a promise on subsequent exports.

In most cases, this would have had no ill effect, since the receiver would know it was a promise based on the first time it was imported. However, if the receiver had dropped the import concurrently with the sender exporting it again, then the receiver may have lost this information.