Fix HashIndex memory leak when doing lots of erasures.
So this is embarassing... the size of a HashIndex would grow more-or-less with the number of insertions, even if most inserted items were removed later.

Fixing this not only fixes the memory leak, but makes the benchmark 30% faster. When compared to unordered_map, HashMap goes from being 66% faster to being 116% faster.