Cleanup compiler macros a bit
Make it easy to test __has_feature without complicated nesting. It's not
globally applicable per se (the comment in common.h remains valid about
some cases needing to distinguish the compiler supporting __has_feature
vs the feature being disabled). In the sanitizer cases though that
restriction can be relaxed because compiler support for a sanitizer is
either available throw __has_feature or a macro (or both), so the
prescence of either is all that's needed. In all other cases you don't
have the feature.