Leak the root exception callback
Avoid any chance for static initialization order subtleties causing
problems if those objects throw during destruction. This way we ensure
that the root exception callback is always available.

There is a proposal in the comment of how this can be done today without
a heap allocation, but I lay out the reason why the intentional leak
seems less complicated with minimal downsides.