Disallow copying of B-Tree table indexes and fix their move operations
Previously, copying of B-Tree table indexes was incorrectly allowed (and
broken) and moving them was also similarly broken.

Specifically, there were two different sources of bad behavior:
* If you copied or moved a B-Tree index, its underlying `tree` memory
  would be freed multiple times.
* And if you assigned one B-Tree index over top of another, the target
  of the assignment would fail to free its underlying `tree` memory.

I found this when debugging a recently-added memory leak involving a
Table with a TreeIndex stored within an `ExternalMutexGuarded`. The
destructor of ExternalMutexGuarded<T> attempts to clear out its
underlying T object while under the protection of the mutex by assigning
a default-constructed T to it. This triggered the bug where we BTreeImpl
failed to free its memory when assigned to.