Clarify that kj's async Timer only updates when the event loop waits
Th original comment was misleading to me. I understood it to mean "Timer counts
on its own, but may get slightly out of sync until the start of the event loop,"
especially because of the "guaranteed to be synchronized" bit. The actual
meaning is "At any given point, Timer is frozen in time, and is only updated at
the start of the event loop."