Disable fibers on FreeBSD, they aren't really supported here.
Most of the related unit-tests fail, which suggests that support
was never really implemented on that platform. So it's better to
mark it as such.

Some extras:

+ Centralize fibers disablement and have an option to still compile
those in by defining KJ_USE_FIBERS=1, as suggested by @vlovich and
@kentonv.

+ Add OpenBSD into list of platforms where fibers are known not to
work.

+ Make KJ_USE_FIBERS and KJ_NO_EXCEPTIONS mutually exclusive, use
KJ_USE_FIBERS to replace KJ_FIBERS_AVAILABLE in test cases.