Work around old OpenSSL versions not being able to verify Let's Encrypt certificate chains.
As of September 30, 2021, old versions of OpenSSL (1.0.x) could no longer verify Let's Encrypt certificate chains by default, due to the problems described here:

https://letsencrypt.org/docs/dst-root-ca-x3-expiration-september-2021/
https://community.letsencrypt.org/t/openssl-client-compatibility-changes-for-let-s-encrypt-certificates/143816

It turns out the problem can be solved by setting a flag. That flag actually became the default in 1.1.0.

Of course, no one should be using OpenSSL 1.0.x anymore, but the various OpenSSL forks fixed this problem at different times, and the FIPS-certified version of BoringSSL is known to have this problem as well, thus requiring the work-around.