Update outdated references to Sandstorm.
Sandstorm-the-company no longer exists, and these references have been inaccurate for many years.

Some places describe Cap'n Proto as being owned by Sandstorm. What this was meant to describe is that Sandstorm was primarily responsible for development at the time, and held copyright of the code it developed. All code is copyright by whoever wrote it, or their respective employer. There are no registered trademarks. Kenton owns the domain, which was registered long before Sandstorm existed.