Add `TlsContext::wrapAddress()` to wrap a `NetworkAddress`.
The functionality already existed internally. I am not really sure why it wasn't exposed. It's useful as `NetworkAddress` is effectively the "connection factory" interface that you pass someone to give them a way to connect to the destination of your choosing, and if you want them to use TLS, you want to give them a `NetworkAddress` with TLS built-in.