Add block quote syntax.
A backtick begins a string literal which includes all text to the end of the line, verbatim, including the newline character. Escape sequences are not processed and there is no way to end the quote except to break the line.

The idea here is to make it easy to write an indented block of verbatim text, without including the indentation in the quote.

Example:

    const myWorkerScript :Text =
        `export default {
        `  async fetch(req, env) {
        `    return new Response("Hello world");
        `  }
        `}