Implement asynchronous environments
This commit adds a new feature to the kj-async library: Environments.

User-facing documentation is available in "A tour of KJ", so this commit message will contain only a few implementation notes.

Our desired semantics for asynchronous environments is for them to act like thread-local storage for the abstract threads of execution formed by promise call graphs. An initial implementation leaned directly on thread-local storage by storing the active environment in the EventLoop data structure, and required every Promise constructor to look up the current EventLoop. This introduced a measurable performance regression for code which did not even use the feature.

The implementation in this commit instead avoids the constant thread-local storage lookups by piggybacking off of the recent PromiseArena optimization. Now, we only need to look up the EventLoop at the start of a new PromiseArena, and the environment set is automatically propagated to all promises which are added to the arena. In practice, these promises are always dependents of the earlier promises in the arena, and this commit requires that this practice become a rule. With this rule in place, we can assume that all promises in an arena are always dependencies of other promises in the same arena (or themself the root which captured the environment). Since all promises created within an environment must be dependents of a common ancestor, or themselves roots, our optimization has semantics which propagate the environment to at least a superset of the promises that our desired semantics would have.

The mismatch between our desired semantics and our optimization's semantics is that it may be possible for environments to propagate to dependent promises that we did not intend. That is, it is impossible for us to statically guarantee that promise arenas do not gain promises which were constructed under a different environment. Instead, we detect this circumstance at runtime in debug mode, and crash (in release mode, it would manifest as a wrong-environment bug). Additionally, we provide a couple tools -- `.adoptEnvironment()` and TaskSets -- to import and export promises across environments.

PromiseNode implementations are responsible for constructing environment scopes wherever necessary to cover user code invocations. This commit adds such scopes to all known user code invocation sites. Some rules to follow to demystify this process for future PromiseNode implementations can be found in the Async Environments comment in async-inl.h.