Optimization: Use static disposer for PromiseNodes.
There are only a couple places where we need dynamic disposers. For now I've introduced a little wrapper to handle these but I'm going to try to remove that in subsequent commits.

This improves performance by a few percent. It also reduces binary size (of http-over-capnp-perf-test) by ~4%, which is not too shabby.