Optimization: Allow `Finish` message to be elided when not needed.
Specifically, when the response contains no capabilities, we can go ahead and clean up the answer table entry early and inform the caller, in the `Return` message, that they don't need to send a `Finish`.

This change doesn't hugely benefit the benchmark, but the benchmark doesn't actually write messages to a socket. IRL, this should save bandwidth on the wire as well as syscalls and packet count.