Fix premature cancellation of streaming calls when capability dropped.
At present, if you make a streaming call on an RPC capability, then drop the capability, the streaming call could be canceled (`Finish` message sent to the server) before it has returned. This is NOT true of regular calls -- keeping the Promise returned by the call alive prevents it from being canceled. The problem happens with streaming calls because the RpcFlowController is dropped when the capability is dropped, and the flow controller is responsible for waiting for streaming calls to complete.

In the case of byte-stream.capnp, it quite normal for the capability to be dropped before streaming is done, due to the fact that KJ streams currently lack an `end()` call, so instead we invoke it in the destructor and then detach the promise. No one was holding on to the capability in this place. But the bug is definitely in the RPC system, not byte-stream: normally, you are allowed to drop a capability even while calls are outstanding, and it is not supposed to cancel those calls.

The reason this didn't cause a problem previously is because `ByteStream::write()` does not allow cancellation, so the cancellation attempts were ignored. However, we'd like to change that, as blocking cancellation is expensive. Unfortunately, we will not be able to roll out a version that allows cancellation until after fully rolling out this bug fix, since all clients must have the fix before the server can safely support cancellation.