Optimization: Avoid ServerRequestContext in http-over-capnp.
This requires introducing a new, alternate RPC method for http-over-capnp. The new method avoids the need for ServerRequestContext by instead relying on `setPipeline()` to set up pipelining before the call has returned. I would have done it this way in the first place, but `setPipeline()` did not yet exist at the time, and I didn't realize it would be easy to add. Oops.

Unfortunately this is a backwards-incompatible change, so will require a multi-stage rollout. HttpOverCapnpFactory will continue using the old protocol until explicitly configured to use the new one. It will support receiving either protocol forever.

This reduces allocations by about 10% in the benchmark, and eliminates a few network messages.