Fix compiler warnings in Alpine 3.17.
This was pushed to Docker Hub a few days ago and it apparently updates to GCC 12, which has new warnings.

Among other things, it complains when `KJ_ASSERT_NONNULL(x)` is used but the result is discarded. This can always be replaced with `KJ_ASSERT(x != nullptr)`, so I did that rather than try to figure out how to squelch the warning.