Fix bug in cancellation blocking.
The detached promise used to make the call un-cancelable did not attach a reference to the LocalClient, because it expected `context` would hold this. But `RpcCallContext` actually doesn't hold a ref to the `ClientHook`, so this wasn't sufficient in the case of RPC calls.

This bug was introduced recently with the `allowCancellation` changes.