HttpServer::drain() should not disrupt a brand-new connection.
For the first request on a connection, if drain() was called before the request headers were fully received, we were unconditionally cancelling the read.

In the case that no bytes at all had been received yet, this would be considered a "clean drain" and the connection could potentially be passed off to a new server (if the application supported this).

But if some bytes had been received, this would necessarily result in the connection being closed without a response, i.e. an error. This situation was pretty unlikely to occur because usually all the headers are received in a single packet, so there's no chance for drain() to happen mid-headers, but it could happen if the headers were large or the connection slow.

Instead of trying to distinguish whether any bytes have been received, this commit assumes that any newly-created connection is created for the purpose of seting a request, and so treats the connection as if bytes had already been received from the start. Hence, we do not stop waiting if we receive `drain()`. We complete the first request, *then* drain.