Fix bug in sendForPipeline().
On the receiving end, we need to make sure the `RpcCallContext` lives at least until `Finish` is received. Otherwise, when the context is destroyed, it will prematurely remove the answer table entry making pipelined calls fail, defeating the purpose.

Unfortunately we'll have to wait for this to roll out fully before we can actually use `sendForPipeline()`.