Fix MessageReader::sizeInWords() double-counting first segment.
AFAICT the main effect of this bug is that RpcSystem::setFlowLimit() would end up enforcing ~half the limit specified (assuming all messages are single-segment). I believe this feature is only used by Sandstorm and the limit value used is fairly high, so this may not have had any real impact.