Fix obscure UAF with sendForPipeline().
`tailPipelinePromise` is returned by `context->onTailCall()`, therefore it must hold a strong ref to the context.

This is hard to trigger, because `tailPipelinePromise` is `exclusiveJoin()`ed with `pipelinePromise`, which itself holds such a ref. So, the only way this promise could still exist after `context` has been dropped is the case where this promise is also imminently going to be canceled. But it is actually possible to wedge a continuation in between that UAFs.

This only seems to kick in with `sendForPipeline()` since in the common case the non-pipeline branch also holds the `context` ref long enough to prevent any issue.