Side fix: disconnect() should set disconnected state earlier.
Previously, when shutting down an RPC connection, we'd first destroy all the objects, then set the state to "disconnected".

This change makes it so we set the state to "disconnected" first, before destroying things.

This is important because some RPC objects have destructors which will send new messages. We don't want to be sending any new messages at this point!

This was making it hard to write the unit test coming in the next commit, because it was triggering weird shutdown behavior.