Fix more subtle bugs with embargos.
The previous fix introduced a problem: If the capability legitimately resolved to a promise capability back on the caller machine, then `whenResolved()` call during `handleDisembargo()` would hang.

In fact, thit was indicative of a deeper mistake: We were not correctly handling the Tribble 4-party race condition at all for pipelined capabilities.

The problem is, pipelined calls are always handled through the `PipelineHook` object on the answer table entry. But, upon return, there was no mechanism to force the PipelineHook to freeze its resolution decisions, as required to resolve the race. I had written some code thta seemed intended to do this, but it only froze the capabilities present in the results message structure. In the common case that the call completed locally, this worked because the `LocalPipeline` would actually end up serving results straight out of this same structure. But, in more complicated scenarios, there was no guarantee that the call's `PipelineHook` was ultimately backed by this copy of the structure. In particular, in the presence of tail calls or `context.setPipeline()`, this couldn't be guaranteed at all.

The new solution is, upon returning from a call, to actually replace the `PipelineHook` with one that follows the rules.