Fix redundant constexpr warning in generated code
Clang 15+ in C++17 mode (or higher) with `-Wdeprecated` issues a warning
in the generated code:

```
warning: out-of-line definition of constexpr static data member is
redundant in C++17 and is deprecated [-Wdeprecated]
```

It is rather annoying for downstream users of Capnp, since there's no
way to disable the specific warning. Instead, fix it here for code
compiled for C++17 and newer.

See [1]:

> A constexpr static data member is implicitly inline and does not need
> to be redeclared at namespace scope. This redeclaration without an
> initializer (formerly required) is still permitted, but is deprecated.
> (since C++17)

[1]: https://en.cppreference.com/w/cpp/language/static