Add RevocableServer, which makes it easy to RPC-wrap unowned objects.
If you have an object you don't own which has some well-defined lifetime, and you want to export it over RPC, until now you had to contend with the fact that the RPC peer could hold on to the capability until long past the underlying object's lifetime, and keep sending RPCs to it. This has been the source of a few UAF bugs and some extremely ugly work-arounds over the years.

RevocableServer is a simple and lightweight way to solve this problem.