Fix kj-tls not working in half-duplex mode.
When EOF was received in the read direction, the write direction would also be treated as disconnected, since they shared the same `disconnected` flag.

I think this flag isn't actually needed at all. It looks like it was intended as an optimization of sorts, to avoid calling OpenSSL if we already knew we were at EOF. But the optimization is wrong, since it doesn't distinguish between read and write directions. So I just removed it.