Define KJ_MSVC_TRADITIONAL_CPP also in Release
The code in common.h for the definition of KJ_MSVC_TRADITIONAL_CPP was
active only in debug. However, it is used in debug.h even when not in
debug. Thus, while the fix worked for debug builds, it did not work
for release builds.

This was probably undetected because CI builds only Debug for MSVC.