http: Call getInner() after every co_await in HttpChunkedEntityWriter
The reference to the inner output stream provided by HttpChunkedEntityWriter's base class's getInner() can be invalidated before the HttpChunkedEntityWriter. To put it another way, body streams which we provide to user code (the HttpChunkedEntityWriter) can be prolonged by the user beyond the lifetime of the underlying connection. getInner() is written in such a way to catch the issue, by checking for validity before returning a reference, and I think the tryPumpFrom() implementation probably should call it after every resumption.