Implement StringPtr::find
I have needed this a few times and always implemented it myself with
memmem or std::search.  Some others have asked for this in #868, so
maybe it’s time we implement this.

This commit contains the interface and a dumb implementation, along
with a few tests.  I’ll try to make the implementation smarter in
later commits.