Replace strstr with StringPtr::find in debug-test
Now that we have StringPtr::contains from #868, we should prefer
StringPtr::contains in preference to calling out to the C standard
library.

I would have liked to replace:

  text.append(what.cStr(), what.cStr()+end);

with:

  text.append(what.slice(0, end));

but slice returns an ArrayPtr rather than a StringPtr and there’s no
clean way of converting an ArrayPtr into a StringPtr.