Optimize: FdObserver uses custom adapted promise.
Previously, FdObserver's events all used newPromiseAndFulfiller<void>(), which requires two allocations. We can reduce this to one by using newAdaptedPromise() instead.

Meanwhile, we can use this custom adapter to keep a count of the number of waiters that currently exist. This is useful for some debug checks and will be useful for more in the next commit.