Refactor: CapnpToKjWebSocketAdapter doesn't depend on RequestState.
Instead, since this is the only remaining user of any of the cancellation functionality, we simply merge it into this class itself.

I think the code ends up somewhat simpler, though I think it could get even moreso if it turns out `disconnect()` isn't really needed. I think this may be the case: WebSocket connections already have an explicit close message and are not intended to run with the socket itself in a half-open state. So maybe we could just remove `disconnect()` entirely and close the underlying socket in the destructor instead. But that's a change that extends outside http-over-capnp, so not doing it here.